Conflicts:
	Sources/Internal/Render/RHI/GLES2/_gl.h

Former-commit-id: eaf181b5c0ec2a496840f5ff796bd7e56e8cbe9b

+++ Sources/Internal/Render/Highlevel/Vegetation/VegetationRenderData.h
@@ -40,7 +40,7 @@
 namespace DAVA
 {
 
-typedef uint16 VegetationIndex;
+typedef uint32 VegetationIndex;
 
 struct VegetationVertex
 {
